pyproject.toml
[build-system]
requires = ["wheel", "cmake", "setuptools>=42", "setuptools_scm[toml]>=3.4"]
build-backend = "setuptools.build_meta"

[project]
name = "praxis"
version = "0.2.0"
description = "for all, none"
authors = [{ name = "Ryan Brooks", email = "ryan@src.eco" }]
urls = { label = "https://src.eco" }
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
    "asciichartpy",
    "attrs",
    "blessed",
    # "bytelatent @ git+https://github.com/facebookresearch/blt@6ffeb66b534e8af62692383fdc49d2c22148ea4a#egg=bytelatent",
    # "blt @ file:///home/crow/repos/blt",
    "cmake",
    "datasets",
    "evaluate",
    "flask",
    "flask-cors",
    "flask-socketio",
    "isort",
    # "lighteval[math]",
    "lightning",
    "lsprotocol",
    "matplotlib",
    "neurallambda @ git+https://github.com/neurallambda/neurallambda",
    "cut-cross-entropy @ git+https://github.com/apple/ml-cross-entropy",
    "networkx",
    "pygls",
    "psutil",
    "pytest",
    "pytorch_optimizer==3.5.1",
    "setuptools",
    "torch",
    "torcheval",
    "transformers",
    "watchdog",
    "pyyaml",
    "werkzeug",
    "wheel",
    "xformers",
    "zstandard",
]

[tool.setuptools]
packages = ["praxis"]

[tool.pytest.ini_options]
addopts = """
    --ignore=tests/test_encoders.py
    --ignore=tests/test_tokenizers.py
    --deselect=tests/test_modeling.py::test_praxis_model_with_encoder_init
    --deselect=tests/test_modeling.py::test_praxis_causal_lm_with_encoder_init
    --deselect=tests/test_modeling.py::test_praxis_model_with_encoder_forward
    --deselect=tests/test_modeling.py::test_praxis_causal_lm_with_encoder_forward
"""
